fix: remove unused provider functions
